Added a full blown test for S3 Storage, backed with testcontainers.
This test, StorageControllerTest runs slow, due to it being backed by LocalStack TestContainer, running a docker image with a local S3. The LocalStackContainer is designed to run within the scope of JUnit. This test, using @SpringBootTest in all its glory, tests our API over http, and therefore needs the LocalStackContainer to be exposed as a Spring bean outside of the JUnit test itself. Say hello to StubbedLocalStackContainer.

(In order to get this to work, it was necessary to exclude android-json from spring-boot-starter-test..*sigh*)